Support !Linux, multiple sensors & autodiscovery
Only Linux has a hidraw driver providing file-based access to HID
devices. For other operating systems such as Windows or macOS, USB HID
devices must be accessed through a library.

Depend on karalabe/hid (with karalabe/hid#13 merged) for USB
communication. This also removes the necessity to manually specify the
correct hidraw device file, since HIDAPI can do enumeration based on the
USB vendor ID and product ID.

This also simplifies supporting multiple devices with a single server
process, so implement that. Note that all CO2 detectors I could get
a hold had non-unique serial numbers, so the device serial number could
not be used to distinguish them in the prometheus output. Instead, the
path of the USB device is used now. Unfortunately this means that the
sensor ID changes when the detector is plugged into a different USB
port.

Also note that these paths can get very large and contain special
characters and spaces, but prometheus does not allow these in gauge
identifiers. To work around this, apply SHA1 to the path and use it as
identifier.